Thermocouple Board Firmware

+++ thermocouple.h
@@ -1,9 +1,10 @@
 #ifndef THERMOCOUPLE_H
 #define THERMOCOUPLE_H
-#include <xc.h>
+
 #include <stdbool.h>
+#include <xc.h>
 
 void init_tc(void);
-void get_tc(uint16_t *data_out, uint8_t tc_num);
+void get_tc(uint32_t *data_out, uint8_t tc_num);
 
 #endif
